chore(core): extracted `computeUseDefinitions` to separate file
