+++ packages/core/src/running/computeRulesWithOptions.ts
@@ -1,15 +1,19 @@
-import { ConfigRuleDefinition } from "../types/configs.js";
 import { AnyRule } from "../types/rules.js";
+import { ConfigUseDefinitionWithFiles } from "./computeUseDefinitions.js";
 
 // TODO: This is very slow and the whole thing should be refactored 🙌.
 // The separate lintFile function recomputes rule options repeatedly.
 // It'd be better to build a collection of groups with options
 export function computeRulesWithOptions(
-	ruleDefinitions: ConfigRuleDefinition[],
+	filePath: string,
+	useDefinitions: ConfigUseDefinitionWithFiles[],
 ) {
+	const ruleDefinitionsToRun = useDefinitions
+		.filter((use) => use.found.has(filePath))
+		.flatMap((use) => use.rules);
 	const rulesWithOptions = new Map<AnyRule, unknown>();
 
-	for (const definition of ruleDefinitions) {
+	for (const definition of ruleDefinitionsToRun) {
 		const [options, rule] =
 			"rule" in definition
 				? [definition.options, definition.rule]
