+++ packages/core/src/running/lintOnce.ts
@@ -1,22 +1,12 @@
-import { makeAbsolute } from "@flint.fyi/utils";
 import { CachedFactory } from "cached-factory";
 import { debugForFile } from "debug-for-file";
-import * as fs from "node:fs/promises";
-import path from "node:path";
 
 import { readFromCache } from "../cache/readFromCache.js";
 import { writeToCache } from "../cache/writeToCache.js";
-import { collectFilesValues } from "../globs/collectFilesValues.js";
-import { AnyLevelDeep } from "../types/arrays.js";
-import {
-	ConfigRuleDefinition,
-	ConfigUseDefinition,
-	ProcessedConfigDefinition,
-} from "../types/configs.js";
-import { FilesGlobObjectProcessed, FilesValue } from "../types/files.js";
+import { ProcessedConfigDefinition } from "../types/configs.js";
 import { AnyLanguage } from "../types/languages.js";
 import { FileResults, LintResults } from "../types/linting.js";
-import { flatten } from "../utils/arrays.js";
+import { computeUseDefinitions } from "./computeUseDefinitions.js";
 import { lintFile } from "./lintFile.js";
 import { readGitignore } from "./readGitignore.js";
 
@@ -31,43 +21,12 @@ export async function lintOnce(
 	configDefinition: ProcessedConfigDefinition,
 	{ ignoreCache, skipDiagnostics }: LintOnceSettings,
 ): Promise<LintResults> {
-	interface ConfigUseDefinitionWithFiles extends ConfigUseDefinition {
-		found: Set<string>;
-		rules: ConfigRuleDefinition[];
-	}
-
 	const gitignore = await readGitignore();
 
 	log("Collecting files from %d use pattern(s)", configDefinition.use.length);
 	log("Excluding based on .gitignore: %s", gitignore);
 
-	const useDefinitions: ConfigUseDefinitionWithFiles[] = await Promise.all(
-		configDefinition.use.map(async (use) => {
-			const globs = resolveUseFilesGlobs(use.files, configDefinition);
-
-			return {
-				...use,
-				found: new Set(
-					(
-						await Array.fromAsync(
-							fs.glob([globs.include].flat(), {
-								exclude: [...gitignore, ...globs.exclude],
-								withFileTypes: true,
-							}),
-						)
-					)
-						.filter((entry) => entry.isFile())
-						.map((entry) =>
-							path.relative(
-								process.cwd(),
-								makeAbsolute(path.join(entry.parentPath, entry.name)),
-							),
-						),
-				),
-				rules: use.rules ? flatten(use.rules) : [],
-			};
-		}),
-	);
+	const useDefinitions = await computeUseDefinitions(configDefinition);
 
 	const allFilePaths = new Set(
 		useDefinitions.flatMap((use) => Array.from(use.found)),
@@ -91,12 +50,10 @@ export async function lintOnce(
 		const { dependencies, diagnostics, reports } =
 			cached?.get(filePath) ??
 			(await lintFile(
-				makeAbsolute(filePath),
+				filePath,
 				languageFactories,
-				useDefinitions
-					.filter((use) => use.found.has(filePath))
-					.flatMap((use) => use.rules),
 				skipDiagnostics,
+				useDefinitions,
 			));
 
 		filesResults.set(filePath, {
@@ -117,41 +74,4 @@ export async function lintOnce(
 	await writeToCache(configDefinition.filePath, lintResults);
 
 	return lintResults;
-}
-
-function collectUseFilesGlobsObject(
-	files: AnyLevelDeep<FilesValue> | undefined,
-	configDefinition: ProcessedConfigDefinition,
-): FilesGlobObjectProcessed {
-	switch (typeof files) {
-		case "function":
-			return resolveUseFilesGlobs(files(configDefinition), configDefinition);
-
-		case "undefined":
-			return { exclude: [], include: [] };
-
-		default: {
-			const exclude = new Set<string>();
-			const include = new Set<string>();
-
-			collectFilesValues(flatten(files), exclude, include);
-
-			return {
-				exclude: Array.from(exclude),
-				include: Array.from(include),
-			};
-		}
-	}
-}
-
-function resolveUseFilesGlobs(
-	files: AnyLevelDeep<FilesValue> | undefined,
-	configDefinition: ProcessedConfigDefinition,
-): FilesGlobObjectProcessed {
-	const globs = collectUseFilesGlobsObject(files, configDefinition);
-
-	return {
-		exclude: [...globs.exclude, ...(configDefinition.ignore ?? [])],
-		include: globs.include,
-	};
 }
